Реорганизация func_break.*
